feat(mdx): ship createMdxComponents() locale-bound factory from package (#2332)
Move the MDX components assembly into the shared package as
`@takazudo/zudo-doc/mdx-components`, exporting
`createMdxComponents({ settings, locale, navData, extras })`. The factory
bakes every package-resident binding — htmlOverrides, admonitions
(Note/Tip/Info/Warning/Danger/Important/Caution), CodeGroup, Tabs/TabItem,
MathBlock, the `img` base-rewrite, and the enlargeable-`p` override — and
threads the active `locale` into the host-supplied `navData` wrappers so
`/ja` resolves the JA collection (the load-bearing locale thread).

Deliberately NOT a static global mdx-components slot: a single static map
cannot be locale-parameterized and would render the EN nav tree on every
`/ja` page. The factory is called once per render with the active locale.

Collapse `pages/_mdx-components.ts` (382 -> 106 lines) to delegate to the
package factory, keeping ONLY the project-bound pieces: the 4 locale-aware
nav wrappers (passed as `navData`) and the showcase stubs + Island /
Details / HtmlPreview / PresetGenerator bindings (passed as `extras`).

Append the `./mdx-components` subpath to the package exports per the S2
JS-subpath-group convention (auto-compiled by the tsup `src/**` glob).

Verified: package build + full `pnpm build` render both `/docs` and
`/ja/docs` (315 pages); locale binding confirmed (JA SiteTreeNav emits
/ja/docs hrefs); `pnpm check` green; 598 package unit tests pass.
