+++ packages/zudo-doc/package.json
@@ -203,6 +203,10 @@
       "types": "./dist/settings.d.ts",
       "default": "./dist/settings.js"
     },
+    "./mdx-components": {
+      "types": "./dist/mdx-components/index.d.ts",
+      "default": "./dist/mdx-components/index.js"
+    },
     "./safelist.css": "./dist/safelist.css",
     "./safelist": "./dist/safelist.css",
     "./content.css": "./dist/content.css",
